[client] Removal of empty directories (after delete and move) made recursive

+++ client/client_daemon.py
@@ -80,14 +80,23 @@ def is_directory(method):
     return wrapper
 
 
-def remove_dir_if_empty(dir_path):
+def remove_dir_if_empty(dir_path, sharing_path, recursive=True):
     """
     Given a directory path, delete it if empty.
+    If <recursive> is True, act recursively on the upper dir.
+
     :param dir_path: str
     """
+    if dir_path == sharing_path:
+        # Do not delete the user sharing folder!
+        return
+
     if not os.listdir(dir_path):
         logger.debug('Removing empty directory {}'.format(dir_path))
         os.rmdir(dir_path)
+
+        if recursive:
+            remove_dir_if_empty(os.path.dirname(dir_path), sharing_path)
 
 
 class Daemon(FileSystemEventHandler):
@@ -362,7 +371,7 @@ class Daemon(FileSystemEventHandler):
             return False
         else:
             # After removing the file, remove the directory if it is empty.
-            remove_dir_if_empty(os.path.dirname(abs_src))
+            remove_dir_if_empty(os.path.dirname(abs_src), self.cfg['sharing_path'])
 
         self.client_snapshot[dst] = self.client_snapshot[src]
         self.client_snapshot.pop(src)
@@ -387,7 +396,7 @@ class Daemon(FileSystemEventHandler):
                            'Error occurred: {}'.format(abs_path, e))
         else:
             # After deleting the file, remove the directory if it is empty.
-            remove_dir_if_empty(os.path.dirname(abs_path))
+            remove_dir_if_empty(os.path.dirname(abs_path), self.cfg['sharing_path'])
 
         if self.client_snapshot.pop(filepath, 'ERROR') != 'ERROR':
             logger.info('Deleted file on client during SYNC.\nDeleted filepath: {}'.format(abs_path))
